refactor(Security): Improve security on put and delete routes.

+++ src/main/java/fr/polytech/service/ReferenceService.java
@@ -1,5 +1,7 @@
 package fr.polytech.service;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.interfaces.DecodedJWT;
 import fr.polytech.model.Reference;
 import fr.polytech.model.ReferenceDTO;
 import fr.polytech.repository.ReferenceRepository;
@@ -7,7 +9,9 @@ import jakarta.ws.rs.NotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.List;
 import java.util.UUID;
@@ -76,17 +80,26 @@ public class ReferenceService {
      * Update a reference.
      *
      * @param reference Reference to update.
+     * @param token     String - Access token.
      * @return Updated reference.
      * @throws NotFoundException If the reference is not found.
      */
-    public Reference updateReference(ReferenceDTO reference) throws NotFoundException {
+    public Reference updateReference(ReferenceDTO reference, String token) throws HttpClientErrorException {
         logger.info("Starting the update of a reference");
         Reference storedReference = referenceRepository.findById(reference.getId()).orElse(null);
 
         if (storedReference == null) {
             logger.error("Error while updating a reference: reference not found");
             // If the reference is not found, throw an exception
-            throw new NotFoundException("Reference not found");
+            throw new HttpClientErrorException(HttpStatus.NOT_FOUND, "Reference not found");
+        }
+
+        boolean isSameUser = checkUser(reference.getSenderId().toString(), token);
+
+        if (!isSameUser) {
+            logger.error("Error while deleting a reference: user not authorized");
+            // If the user is not authorized, throw an exception
+            throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED, "User not authorized");
         }
 
         storedReference.setMessage(reference.getMessage());
@@ -100,19 +113,43 @@ public class ReferenceService {
     /**
      * Delete a reference.
      *
-     * @param id Reference id.
+     * @param id    Reference id.
+     * @param token String - Access token.
      * @throws NotFoundException If the reference is not found.
      */
-    public void deleteReference(UUID id) throws NotFoundException {
+    public void deleteReference(UUID id, String token) throws HttpClientErrorException {
         logger.info("Starting the deletion of a reference");
         Reference reference = referenceRepository.findById(id).orElse(null);
 
         if (reference == null) {
             logger.error("Error while deleting a reference: reference not found");
             // If the reference is not found, throw an exception
-            throw new NotFoundException("Reference not found");
+            throw new HttpClientErrorException(HttpStatus.NOT_FOUND, "Reference not found");
+        }
+
+        boolean isSameUser = checkUser(reference.getSenderId().toString(), token);
+
+        if (!isSameUser) {
+            logger.error("Error while deleting a reference: user not authorized");
+            // If the user is not authorized, throw an exception
+            throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED, "User not authorized");
         }
 
         referenceRepository.delete(reference);
+    }
+
+    /**
+     * Check if the "sub" field of the access token matches the user id
+     *
+     * @param id    String - User id
+     * @param token String - Access token
+     * @return boolean - True if the "sub" field of the access token matches the user id, false otherwise
+     * @throws HttpClientErrorException if an error occurs while decoding the token
+     */
+    public boolean checkUser(String id, String token) throws HttpClientErrorException {
+        String actualToken = token.split("Bearer ")[1];
+        DecodedJWT jwt = JWT.decode(actualToken);
+        String userId = jwt.getClaim("sub").asString();
+        return userId.equals(id);
     }
 }
